src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/HTMLTemplateElement/cycles-in-shadow-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: div.createShadowRoot is not a function
Test that cycle detection traverses over both templates and shadow roots

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
